fixed error traps on customers and branches as per Ricards report
git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@6548 09516a42-71f9-43be-811f-9b821cf0d246

+++ CustomerBranches.php
@@ -53,7 +53,7 @@ if (isset($_POST['submit'])) {
 	if ($_SESSION['SalesmanLogin'] != '') {
 		$_POST['Salesman'] = $_SESSION['SalesmanLogin'];
 	}
-	if (ContainsIllegalCharacters($_POST['BranchCode']) OR mb_strstr($_POST['BranchCode'],' ') OR mb_strstr($_POST['BranchCode'],'-')) {
+	if (ContainsIllegalCharacters($_POST['BranchCode']) OR mb_strstr($_POST['BranchCode'],' ')) {
 		$InputError = 1;
 		prnMsg(_('The Branch code cannot contain any of the following characters')." -  &amp; \' &lt; &gt;",'error');
 		$Errors[$i] = 'BranchCode';
@@ -340,14 +340,24 @@ if (isset($_POST['submit'])) {
 						prnMsg(_('Cannot delete this branch because contract have been created that refer to it') . '. ' . _('Purge old contracts first'),'warn');
 						echo '<br />' . _('There are') . ' ' . $myrow[0] . ' '._('contracts referring to this branch/customer');
 					} else {
-						$SQL="DELETE FROM custbranch WHERE branchcode='" . $SelectedBranch . "' AND debtorno='" . $DebtorNo . "'";
-						if ($_SESSION['SalesmanLogin'] != '') {
-							$SQL .= " AND custbranch.salesman='" . $_SESSION['SalesmanLogin'] . "'";
-						}
-						$ErrMsg = _('The branch record could not be deleted') . ' - ' . _('the SQL server returned the following message');
-						$result = DB_query($SQL,$db,$ErrMsg);
-						if (DB_error_no($db)==0){
-							prnMsg(_('Branch Deleted'),'success');
+						//check if this it the last customer branch - don't allow deletion of the last branch
+						$SQL = "SELECT COUNT(*) FROM custbranch WHERE debtorno='" . $DebtorNo . "'";
+
+						$result = DB_query($SQL,$db);
+						$myrow = DB_fetch_row($result);
+	
+						if ($myrow[0]==1) {
+							prnMsg(_('Cannot delete this branch because it is the only branch defined for this customer.'),'warn');
+						} else {
+							$SQL="DELETE FROM custbranch WHERE branchcode='" . $SelectedBranch . "' AND debtorno='" . $DebtorNo . "'";
+							if ($_SESSION['SalesmanLogin'] != '') {
+								$SQL .= " AND custbranch.salesman='" . $_SESSION['SalesmanLogin'] . "'";
+							}
+							$ErrMsg = _('The branch record could not be deleted') . ' - ' . _('the SQL server returned the following message');
+							$result = DB_query($SQL,$db,$ErrMsg);
+							if (DB_error_no($db)==0){
+								prnMsg(_('Branch Deleted'),'success');
+							}
 						}
 					}
 				}
@@ -394,16 +404,16 @@ if (!isset($SelectedBranch)){
 				 ' . ' ' . _('Branches defined for'). ' '. $DebtorNo . ' - ' . $myrow[0] . '</p>';
 		echo '<table class="selection">
 			<tr>
-				<th>' . _('Code') . '</th>
-				<th>' . _('Name') . '</th>
-				<th>' . _('Branch Contact') . '</th>
-				<th>' . _('Salesman') . '</th>
-				<th>' . _('Area') . '</th>
-				<th>' . _('Phone No') . '</th>
-				<th>' . _('Fax No') . '</th>
-				<th>' . _('Email') . '</th>
-				<th>' . _('Tax Group') . '</th>
-				<th>' . _('Enabled?') . '</th>
+				<th class="ascending">' . _('Code') . '</th>
+				<th class="ascending">' . _('Name') . '</th>
+				<th class="ascending">' . _('Branch Contact') . '</th>
+				<th class="ascending">' . _('Salesman') . '</th>
+				<th class="ascending">' . _('Area') . '</th>
+				<th class="ascending">' . _('Phone No') . '</th>
+				<th class="ascending">' . _('Fax No') . '</th>
+				<th class="ascending">' . _('Email') . '</th>
+				<th class="ascending">' . _('Tax Group') . '</th>
+				<th class="ascending">' . _('Enabled?') . '</th>
 			</tr>';
 
 		$k=0;
